Split logic into fake database and fake api calls
